Crear bases y tablas de conteo y tasas estandarizadas de mortalidad

API para la generación de bases resumen para la creación de gráficos de tendencia y tablas reporte con información nominal (por sujeto).

cdc_mortality(data, denominador, estandar, ...,
  rute_name = NA_character_)

cdc_denominator(denominador, estrato, ...)

cdc_mortality_2(data, denominador, estandar, ...,
  rute_name = NA_character_)

cdc_mortality_adj(data, denominador, estandar, ...,
  rute_name = NA_character_)

cdc_mortality_raw(data, denominador, ..., rute_name = NA_character_)

cdc_mortality_to_wide(data_long, rute_name = NA)

Arguments

data
data frame con información nominal (por sujeto)

denominador
data frame con cantidad estimada de personas por rango de edad, sexo, departamento y año

estandar
data frame con proporciones otorgadas por la OMS para estandarizar tasas (fuente: https://seer.cancer.gov/stdpopulations/world.who.html)

...
lista de una o más variables

rute_name
character string con la ruta y nombre del objeto a exportar. Las bases se exportan en RDS y DTA. Las tablas en CSV y XLSX.

estrato
estrato de edad a usar para estandarizar

data_long
output from cdc_mortality

Value

Base con conteo, población expuesta, tasas crudas y estandarizadas.

Functions

cdc_mortality:

cdc_denominator: create denominator

cdc_mortality_2: create mortality with extra variables

cdc_mortality_adj: create mortality with extra variables

cdc_mortality_raw: create mortality with extra variables

cdc_mortality_to_wide: create output table

Examples

# dependencies ------------------------------------------------------------

#flexibility on dot-dot-dot
#https://stackoverflow.com/questions/7028385/can-i-remove-an-element-in-dot-dot-dot-and-pass-it-on
#https://www.burns-stat.com/the-three-dots-construct-in-r/

library(cdcper)
library(tidyverse)
library(magrittr)
library(readxl)
library(haven)
library(naniar)
library(ggrepel)
library(labelled)
library(gridExtra)

# importar base ----------------------------------------------------------------

data(mortdb)
mortdb

#> # A tibble: 2,258 x 30
#>    cod_sub_cate sexo  avpp_n descripcion edad_a edadasis edadesco edadmef
#>    <chr>        <fct>  <dbl> <chr>        <dbl> <fct>    <fct>    <fct>  
#>  1 K746         Mujer   3.23 OTRAS CIRR~ 71     Adulto ~ No esco~ No per~
#>  2 J189         Mujer   0    NEUMONIA, ~ 95     Adulto ~ No esco~ No per~
#>  3 J189         Homb~  74.1  NEUMONIA, ~  0.167 Infancia No esco~ No per~
#>  4 J189         Homb~  28.2  NEUMONIA, ~ 46     Adulto   No esco~ No per~
#>  5 J841         Homb~  13.2  OTRAS ENFE~ 61     Adulto ~ No esco~ No per~
#>  6 E146         Homb~  18.2  DIABETES M~ 56     Adulto   No esco~ No per~
#>  7 J189         Mujer  32.2  NEUMONIA, ~ 42     Adulto   No esco~ Period~
#>  8 J470         Mujer   0    BRONQUIECT~ 80     Adulto ~ No esco~ No per~
#>  9 I219         Homb~  14.2  INFARTO AG~ 60     Adulto ~ No esco~ No per~
#> 10 C910         Mujer   2.23 LEUCEMIA L~ 72     Adulto ~ No esco~ No per~
#> # ... with 2,248 more rows, and 22 more variables: edaddgsp <fct>,
#> #   descrip_grupo110 <fct>, descrip_grupo10 <fct>, edad_q <chr>, mesfall <fct>,
#> #   year <chr>, edad_q2 <fct>, cie10 <chr>, quinquenal <fct>, trienal <fct>,
#> #   bienal <fct>, departamento <fct>, provincia <chr>, distrito <chr>,
#> #   qpobreza <fct>, condpobrez <fct>, region_nat <fct>, ambito <fct>,
#> #   ubigeo <chr>, edad_asis2015 <chr>, edad_asis_muna <chr>,
#> #   edad_estandar <fct>

# importar denominadores -------------------------------------------------------

data(denominadores_departamento_2000_2020_peru)
denomi_pe <- denominadores_departamento_2000_2020_peru

data(edad_estandarizada_who)
fctwho_std <- edad_estandarizada_who

# CDR + DAR ---------------------------------------------

# formato de base nominal
mortdb %>% glimpse()

#> Observations: 2,258
#> Variables: 30
#> $ cod_sub_cate     <chr> "K746", "J189", "J189", "J189", "J841", "E146", "J...
#> $ sexo             <fct> Mujer, Mujer, Hombre, Hombre, Hombre, Hombre, Muje...
#> $ avpp_n           <dbl> 3.2300, 0.0000, 74.0633, 28.2300, 13.2300, 18.2300...
#> $ descripcion      <chr> "OTRAS CIRROSIS DEL HIGADO Y LAS NO ESPECIFICADAS"...
#> $ edad_a           <dbl> 71.0000, 95.0000, 0.1667, 46.0000, 61.0000, 56.000...
#> $ edadasis         <fct> Adulto Mayor, Adulto Mayor, Infancia, Adulto, Adul...
#> $ edadesco         <fct> No escolar, No escolar, No escolar, No escolar, No...
#> $ edadmef          <fct> No periodo reproductivo, No periodo reproductivo, ...
#> $ edaddgsp         <fct> Adulto Mayor, Adulto Mayor, Niño, Adulto, Adulto M...
#> $ descrip_grupo110 <fct> "Cirrosis y ciertas otras enfermedades crónicas de...
#> $ descrip_grupo10  <fct> Enfermedades del sistema digestivo, Enfermedades I...
#> $ edad_q           <chr> "70-74", "80-+", "<1", "45-49", "60-64", "55-59", ...
#> $ mesfall          <fct> Febrero, Marzo, Setiembre, Setiembre, Abril, Agost...
#> $ year             <chr> "2013", "2013", "2013", "2013", "2013", "2013", "2...
#> $ edad_q2          <fct> 70-74, 80+, 0-4, 45-49, 60-64, 55-59, 40-44, 80+, ...
#> $ cie10            <chr> "K746", "J189", "J189", "J189", "J841", "E146", "J...
#> $ quinquenal       <fct> 2011-2015, 2011-2015, 2011-2015, 2011-2015, 2011-2...
#> $ trienal          <fct> 2013-2015, 2013-2015, 2013-2015, 2013-2015, 2013-2...
#> $ bienal           <fct> 2012-2013, 2012-2013, 2012-2013, 2012-2013, 2012-2...
#> $ departamento     <fct> JUNIN, LIMA, LORETO, AMAZONAS, LAMBAYEQUE, TACNA, ...
#> $ provincia        <chr> "HUANCAYO", "LIMA", "MARISCAL RAMON CASTILLA", "BA...
#> $ distrito         <chr> "CHILCA", "LIMA", "RAMON CASTILLA", "IMAZA", "POMA...
#> $ qpobreza         <fct> Pobre, Aceptable, Muy pobre, Pobre, Regular, Regul...
#> $ condpobrez       <fct> Pobre, No pobre, Pobre, Pobre, No pobre, No pobre,...
#> $ region_nat       <fct> Sierra, Costa, Selva, Selva, Costa, Costa, Costa, ...
#> $ ambito           <fct> Urbano, Urbano, Rural, Rural, Urbano, Urbano, Urba...
#> $ ubigeo           <chr> "120107", "150101", "160401", "010205", "140118", ...
#> $ edad_asis2015    <chr> "60a>", "60a>", "0-11a", "30-59a", "60a>", "30-59a...
#> $ edad_asis_muna   <chr> "60a>", "60a>", "0-9a", "20-59a", "60a>", "20-59a"...
#> $ edad_estandar    <fct> 70-74, 80+, 0-4, 45-49, 60-64, 55-59, 40-44, 80+, ...

#cuando mort y deno comparten covariables
cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year)

#> Joining, by = c("year", "edad_estandar")

#> Joining, by = "edad_estandar"

#> # A tibble: 3 x 6
#>   country year  n_sum  pop_sum cdr_cmil dar_cmil
#>   <chr>   <chr> <int>    <dbl>    <dbl>    <dbl>
#> 1 PERU    2013    729 30475144     2.39     2.92
#> 2 PERU    2014    778 30814175     2.52     3.02
#> 3 PERU    2015    751 31151643     2.41     2.83

cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year,sexo)

#> Joining, by = c("year", "sexo", "edad_estandar")

#> Warning: Column `sexo` joining factor and character vector, coercing into character vector

#> Joining, by = "edad_estandar"

#> # A tibble: 6 x 7
#>   country year  sexo   n_sum  pop_sum cdr_cmil dar_cmil
#>   <chr>   <chr> <chr>  <int>    <dbl>    <dbl>    <dbl>
#> 1 PERU    2013  Hombre   385 15271062     2.52     3.27
#> 2 PERU    2013  Mujer    344 15204082     2.26     2.60
#> 3 PERU    2014  Hombre   410 15438887     2.66     3.38
#> 4 PERU    2014  Mujer    368 15375288     2.39     2.69
#> 5 PERU    2015  Hombre   417 15605814     2.67     3.35
#> 6 PERU    2015  Mujer    334 15545829     2.15     2.36

cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year,sexo,departamento)

#> Joining, by = c("year", "sexo", "departamento", "edad_estandar")

#> Warning: Column `sexo` joining factor and character vector, coercing into character vector

#> Warning: Column `departamento` joining factor and character vector, coercing into character vector

#> Joining, by = "edad_estandar"

#> # A tibble: 150 x 8
#>    country year  sexo   departamento n_sum pop_sum cdr_cmil dar_cmil
#>    <chr>   <chr> <chr>  <chr>        <int>   <dbl>    <dbl>    <dbl>
#>  1 PERU    2013  Hombre AMAZONAS         7  220810     3.17     4.40
#>  2 PERU    2013  Hombre ANCASH          13  575331     2.26     2.87
#>  3 PERU    2013  Hombre APURIMAC         6  231664     2.59     3.90
#>  4 PERU    2013  Hombre AREQUIPA        22  625246     3.52     4.01
#>  5 PERU    2013  Hombre AYACUCHO         8  343265     2.33     3.50
#>  6 PERU    2013  Hombre CAJAMARCA       22  765045     2.88     4.12
#>  7 PERU    2013  Hombre CALLAO          20  490081     4.08     4.48
#>  8 PERU    2013  Hombre CUSCO           16  659633     2.43     3.04
#>  9 PERU    2013  Hombre HUANCAVELICA     7  244132     2.87     3.23
#> 10 PERU    2013  Hombre HUANUCO         11  430580     2.55     4.02
#> # ... with 140 more rows

cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year,departamento)

#> Joining, by = c("year", "departamento", "edad_estandar")

#> Warning: Column `departamento` joining factor and character vector, coercing into character vector

#> Joining, by = "edad_estandar"

#> # A tibble: 75 x 7
#>    country year  departamento n_sum pop_sum cdr_cmil dar_cmil
#>    <chr>   <chr> <chr>        <int>   <dbl>    <dbl>    <dbl>
#>  1 PERU    2013  AMAZONAS        13  419404     3.10     4.66
#>  2 PERU    2013  ANCASH          26 1135962     2.29     2.61
#>  3 PERU    2013  APURIMAC        10  454324     2.20     2.58
#>  4 PERU    2013  AREQUIPA        39 1259162     3.10     3.36
#>  5 PERU    2013  AYACUCHO        20  673609     2.97     4.28
#>  6 PERU    2013  CAJAMARCA       47 1519764     3.09     4.56
#>  7 PERU    2013  CALLAO          31  982800     3.15     3.41
#>  8 PERU    2013  CUSCO           42 1300609     3.23     3.98
#>  9 PERU    2013  HUANCAVELICA    10  487472     2.05     2.50
#> 10 PERU    2013  HUANUCO         16  847714     1.89     2.59
#> # ... with 65 more rows

cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std, year)

#> Joining, by = c("year", "edad_estandar")

#> Joining, by = "edad_estandar"

#> # A tibble: 3 x 6
#>   country year  n_sum  pop_sum cdr_cmil dar_cmil
#>   <chr>   <chr> <int>    <dbl>    <dbl>    <dbl>
#> 1 PERU    2013    729 30475144     2.39     2.92
#> 2 PERU    2014    778 30814175     2.52     3.02
#> 3 PERU    2015    751 31151643     2.41     2.83

#cuando mort y deno no comparten covariables
deno_ys <- cdc_denominator(denominador = denomi_pe,year,sexo)
cdc_mortality_2(data = mortdb,
                denominador = deno_ys,
                estandar = fctwho_std,
                year,sexo,descrip_grupo10)

#> Joining, by = c("year", "sexo")

#> Warning: Column `sexo` joining factor and character vector, coercing into character vector

#> Joining, by = "edad_estandar"

#> # A tibble: 60 x 8
#>    country year  sexo   descrip_grupo10          n_sum pop_sum cdr_cmil dar_cmil
#>    <chr>   <chr> <chr>  <fct>                    <int>   <dbl>    <dbl>    <dbl>
#>  1 PERU    2013  Hombre Enfermedades Infecciosa~    67  2.60e8  0.0258   0.0168 
#>  2 PERU    2013  Hombre Enfermedades neoplasicas    69  2.60e8  0.0266   0.0151 
#>  3 PERU    2013  Hombre Enfermedades del aparat~    68  2.60e8  0.0262   0.0139 
#>  4 PERU    2013  Hombre Afecciones perinatales      11  2.60e8  0.00424  0.00638
#>  5 PERU    2013  Hombre Lesiones y causas exter~    69  2.60e8  0.0266   0.0289 
#>  6 PERU    2013  Hombre Enfermedades mentales y~     7  2.60e8  0.00270  0.00209
#>  7 PERU    2013  Hombre Enfermedades del sistem~    24  2.60e8  0.00924  0.00481
#>  8 PERU    2013  Hombre Enfermedades del sistem~    33  2.60e8  0.0127   0.00732
#>  9 PERU    2013  Hombre Enfermedades metabolica~    20  2.60e8  0.00770  0.00523
#> 10 PERU    2013  Hombre Demas enfermedades          17  2.60e8  0.00655  0.00462
#> # ... with 50 more rows

#o todo en una
cdc_mortality_2(data = mortdb,
                denominador = cdc_denominator(denominador = denomi_pe,year),
                estandar = fctwho_std,
                year,edad_q)

#> Joining, by = "year"

#> Joining, by = "edad_estandar"

#> # A tibble: 54 x 7
#>    country year  edad_q n_sum   pop_sum cdr_cmil dar_cmil
#>    <chr>   <chr> <chr>  <int>     <dbl>    <dbl>    <dbl>
#>  1 PERU    2013  <1        43 518077448 0.00830   0.0125 
#>  2 PERU    2013  01-04     20 518077448 0.00386   0.00581
#>  3 PERU    2013  05-09      7 518077448 0.00135   0.00200
#>  4 PERU    2013  10-14      9 518077448 0.00174   0.00254
#>  5 PERU    2013  15-19      5 518077448 0.000965  0.00139
#>  6 PERU    2013  20-24     16 518077448 0.00309   0.00431
#>  7 PERU    2013  25-29     11 518077448 0.00212   0.00286
#>  8 PERU    2013  30-34     19 518077448 0.00367   0.00474
#>  9 PERU    2013  35-39     21 518077448 0.00405   0.00493
#> 10 PERU    2013  40-44     22 518077448 0.00425   0.00476
#> # ... with 44 more rows

cdc_mortality_2(data = mortdb %>% filter(edad_estandar=="0-4"),
                denominador = cdc_denominator(denominador = denomi_pe,year),
                estandar = fctwho_std,
                year,edad_estandar)

#> Joining, by = "year"

#> Joining, by = "edad_estandar"

#> # A tibble: 51 x 7
#>    country year  edad_estandar n_sum  pop_sum cdr_cmil dar_cmil
#>    <chr>   <chr> <fct>         <int>    <dbl>    <dbl>    <dbl>
#>  1 PERU    2013  0-4              63 30475144    0.207   0.0183
#>  2 PERU    2013  5-9               0 30475144    0       0     
#>  3 PERU    2013  10-14             0 30475144    0       0     
#>  4 PERU    2013  15-19             0 30475144    0       0     
#>  5 PERU    2013  20-24             0 30475144    0       0     
#>  6 PERU    2013  25-29             0 30475144    0       0     
#>  7 PERU    2013  30-34             0 30475144    0       0     
#>  8 PERU    2013  35-39             0 30475144    0       0     
#>  9 PERU    2013  40-44             0 30475144    0       0     
#> 10 PERU    2013  45-49             0 30475144    0       0     
#> # ... with 41 more rows

#cuando no se requiere tasas
mortdb %>% count(year,edadmef) #edad_fertil

#> # A tibble: 6 x 3
#>   year  edadmef                     n
#>   <chr> <fct>                   <int>
#> 1 2013  Periodo Reproductivo       46
#> 2 2013  No periodo reproductivo   683
#> 3 2014  Periodo Reproductivo       61
#> 4 2014  No periodo reproductivo   717
#> 5 2015  Periodo Reproductivo       51
#> 6 2015  No periodo reproductivo   700

mortdb %>% count(year,edad_asis2015) #etapas de vida

#> # A tibble: 15 x 3
#>    year  edad_asis2015     n
#>    <chr> <chr>         <int>
#>  1 2013  0-11a            73
#>  2 2013  12-17a            9
#>  3 2013  18-29a           29
#>  4 2013  30-59a          157
#>  5 2013  60a>            461
#>  6 2014  0-11a            77
#>  7 2014  12-17a           12
#>  8 2014  18-29a           39
#>  9 2014  30-59a          155
#> 10 2014  60a>            495
#> 11 2015  0-11a            78
#> 12 2015  12-17a            8
#> 13 2015  18-29a           29
#> 14 2015  30-59a          161
#> 15 2015  60a>            475

# opcionalmente, se puede exportar en RDS y DTA
# cdc_mortality(data = mortdb,denominador = denomi_pe,estandar = fctwho_std,year,sexo,
#               rute_name = "data/mortalidad_anual_nacional_sexo")

# formato wide para exportar tabla resumen en CSV y XLSX
mortdb %>%
  cdc_mortality(denominador = denomi_pe,estandar = fctwho_std,year,sexo,departamento) %>%
  cdc_mortality_to_wide()

#> Joining, by = c("year", "sexo", "departamento", "edad_estandar")

#> Warning: Column `sexo` joining factor and character vector, coercing into character vector

#> Warning: Column `departamento` joining factor and character vector, coercing into character vector

#> Joining, by = "edad_estandar"

#> # A tibble: 200 x 7
#>    country sexo   departamento key         `2013`    `2014`    `2015`
#>    <chr>   <chr>  <chr>        <fct>        <dbl>     <dbl>     <dbl>
#>  1 PERU    Hombre AMAZONAS     n_sum         7         7         9   
#>  2 PERU    Hombre AMAZONAS     pop_sum  220810    221729    222536   
#>  3 PERU    Hombre AMAZONAS     cdr_cmil      3.17      3.16      4.04
#>  4 PERU    Hombre AMAZONAS     dar_cmil      4.40      3.72      4.67
#>  5 PERU    Hombre ANCASH       n_sum        13        21        16   
#>  6 PERU    Hombre ANCASH       pop_sum  575331    579214    582978   
#>  7 PERU    Hombre ANCASH       cdr_cmil      2.26      3.63      2.74
#>  8 PERU    Hombre ANCASH       dar_cmil      2.87      4.26      3.23
#>  9 PERU    Hombre APURIMAC     n_sum         6         9         4   
#> 10 PERU    Hombre APURIMAC     pop_sum  231664    232984    234224   
#> # ... with 190 more rows

# mortdb %>%
#   cdc_mortality(denominador = denomi_pe,estandar = fctwho_std,year,sexo,departamento) %>%
#   cdc_mortality_to_wide(rute_name = "table/mortalidad_anual_nacional_sexo_departamento")
#