Descarga de información meteorológica

Funciones que permiten la descarga de información meteorológica (v0.2.1)

get_data_gri(var, dep = NULL, pro = NULL, dis = NULL, ids = NULL,
  fuente = NULL, ruta = NULL, rango = NULL, esc_temp = NULL,
  down = FALSE, filename = paste0("Grilla_", var, ".csv"))

get_ids(fuente = "local", ruta = NULL)

var_fun(x, var)

Arguments

var
Variable que será buscada. Puede ser "Prec", "Tmax" o "Tmin".

dep
Nombre o código (en caracter) del departamento.

pro
Nombre o código (en caracter) de la provincia.

dis
Nombre o código (en caracter) del distrito.

ids
Objeto que contiene la metadata.

fuente
Ubicación de la base de datos. Esta puede ser "local" o "web".

ruta
Dirección de la carpeta que contiene los datos. Solo funciona cuando fuente = "local".

rango
Fecha inicial y final que será filtrada.

esc_temp
Escala temporal en la que se descargará la información. Puede ser "Semana", "Mes", "Año" o NULL (vendría a ser escala diaria).

down
Indique si la información se descargará en un archivo (TRUE) o no (FALSE).

filename
Nombre del archivo donde se exportará la información.

x
Datos serán sumados (precipitación) o promediados (el resto).

var
Nombre de la variable

Value

Tabla con información de una variable meteorológica para una lugar, tiempo y escala temporal específica.

Details

Usos

- La información debe ser utilizada de manera referencial y no debe ser considerado como el dato real.

- Utilizar la información para asociar el comportamiento temporal del clima con alguna variable a estudiar.

- La información debe ser utilizada como representativa del área de un distrito y no de una ubicación específica.

- La información puede ser utilizada para determinar el comportamiento espacial del clima entre distritos

- La información puede ser utilizada para analizar eventos extremos, mas no para determinar el valor numérico del evento.

Functions

get_data_gri: función que realiza la descarga de información

get_ids: función que contiene la ubicación de la información

var_fun: var_fun (description pending)

Examples

library(dplyr)
library(googledrive)

# Por primera vez: poner 1 -> ir a correo -> aceptar
ids.web <- get_ids(fuente = "web")

#> Using an auto-discovered, cached token.
#> To suppress this message, modify your code or options to clearly consent to the use of a cached token.
#> See gargle's "Non-interactive auth" vignette for more details:
#> https://gargle.r-lib.org/articles/non-interactive-auth.html
#> The googledrive package is using a cached token for andree.valle@upch.pe.

#> Auto-refreshing stale OAuth token.

#> File downloaded:
#>   * meta_grilla_web.csv
#> Saved locally as:
#>   * C:\Users\AVALLE~1\AppData\Local\Temp\Rtmpee6xvT\file271c2f2a26be

prec <- get_data_gri(
   var = "Prec",
          dep = "LORETO",
   pro = "MAYNAS",
   dis = "IQUITOS",
   fuente = "web",
   ids = ids.web,
   esc_temp = "Mes",
   rango = c("2010-01-01","2018-12-31"))

#> File downloaded:
#>   * 160101.csv
#> Saved locally as:
#>   * C:\Users\AVALLE~1\AppData\Local\Temp\Rtmpee6xvT\file271c3fe4661a

# final product
prec

#> # A tibble: 108 x 7
#> # Groups:   Year [9]
#>     Year Month Fecha       Prec Departamento Provincia Distrito
#>    <int> <int> <date>     <dbl> <chr>        <chr>     <chr>   
#>  1  2010     1 2010-01-01 306.  LORETO       MAYNAS    IQUITOS 
#>  2  2010     2 2010-02-01 283.  LORETO       MAYNAS    IQUITOS 
#>  3  2010     3 2010-03-01 463.  LORETO       MAYNAS    IQUITOS 
#>  4  2010     4 2010-04-01 174.  LORETO       MAYNAS    IQUITOS 
#>  5  2010     5 2010-05-01 240.  LORETO       MAYNAS    IQUITOS 
#>  6  2010     6 2010-06-01 192.  LORETO       MAYNAS    IQUITOS 
#>  7  2010     7 2010-07-01 188.  LORETO       MAYNAS    IQUITOS 
#>  8  2010     8 2010-08-01  71.4 LORETO       MAYNAS    IQUITOS 
#>  9  2010     9 2010-09-01 116.  LORETO       MAYNAS    IQUITOS 
#> 10  2010    10 2010-10-01 132.  LORETO       MAYNAS    IQUITOS 
#> # ... with 98 more rows