CDC databases: Mortality, Hospital outflows, and more

Random subset of mortality (2013-2015), hospital outflows (2000-2015) databases from Peru

data(mortdb)

denominadores_departamento_2000_2020_peru

edad_estandarizada_who

egresos_hospitalarios

Format

An object of class dataset.

Source

CDC-MINSA, <https://www.dge.gob.pe/portal/> ; World (WHO 2000-2025) Standard <https://seer.cancer.gov/stdpopulations/world.who.html>

References

CDC-MINSA ([dge.gob.pe](https://www.dge.gob.pe/portal/docs/asis/Asis_mortdb.pdf))

Examples

library(cdcper)
library(tidyverse)
data(mortdb)
mortdb %>% glimpse()

#> Observations: 2,258
#> Variables: 30
#> $ cod_sub_cate     <chr> "K746", "J189", "J189", "J189", "J841", "E146", "J...
#> $ sexo             <fct> Mujer, Mujer, Hombre, Hombre, Hombre, Hombre, Muje...
#> $ avpp_n           <dbl> 3.2300, 0.0000, 74.0633, 28.2300, 13.2300, 18.2300...
#> $ descripcion      <chr> "OTRAS CIRROSIS DEL HIGADO Y LAS NO ESPECIFICADAS"...
#> $ edad_a           <dbl> 71.0000, 95.0000, 0.1667, 46.0000, 61.0000, 56.000...
#> $ edadasis         <fct> Adulto Mayor, Adulto Mayor, Infancia, Adulto, Adul...
#> $ edadesco         <fct> No escolar, No escolar, No escolar, No escolar, No...
#> $ edadmef          <fct> No periodo reproductivo, No periodo reproductivo, ...
#> $ edaddgsp         <fct> Adulto Mayor, Adulto Mayor, Niño, Adulto, Adulto M...
#> $ descrip_grupo110 <fct> "Cirrosis y ciertas otras enfermedades crónicas de...
#> $ descrip_grupo10  <fct> Enfermedades del sistema digestivo, Enfermedades I...
#> $ edad_q           <chr> "70-74", "80-+", "<1", "45-49", "60-64", "55-59", ...
#> $ mesfall          <fct> Febrero, Marzo, Setiembre, Setiembre, Abril, Agost...
#> $ year             <chr> "2013", "2013", "2013", "2013", "2013", "2013", "2...
#> $ edad_q2          <fct> 70-74, 80+, 0-4, 45-49, 60-64, 55-59, 40-44, 80+, ...
#> $ cie10            <chr> "K746", "J189", "J189", "J189", "J841", "E146", "J...
#> $ quinquenal       <fct> 2011-2015, 2011-2015, 2011-2015, 2011-2015, 2011-2...
#> $ trienal          <fct> 2013-2015, 2013-2015, 2013-2015, 2013-2015, 2013-2...
#> $ bienal           <fct> 2012-2013, 2012-2013, 2012-2013, 2012-2013, 2012-2...
#> $ departamento     <fct> JUNIN, LIMA, LORETO, AMAZONAS, LAMBAYEQUE, TACNA, ...
#> $ provincia        <chr> "HUANCAYO", "LIMA", "MARISCAL RAMON CASTILLA", "BA...
#> $ distrito         <chr> "CHILCA", "LIMA", "RAMON CASTILLA", "IMAZA", "POMA...
#> $ qpobreza         <fct> Pobre, Aceptable, Muy pobre, Pobre, Regular, Regul...
#> $ condpobrez       <fct> Pobre, No pobre, Pobre, Pobre, No pobre, No pobre,...
#> $ region_nat       <fct> Sierra, Costa, Selva, Selva, Costa, Costa, Costa, ...
#> $ ambito           <fct> Urbano, Urbano, Rural, Rural, Urbano, Urbano, Urba...
#> $ ubigeo           <chr> "120107", "150101", "160401", "010205", "140118", ...
#> $ edad_asis2015    <chr> "60a>", "60a>", "0-11a", "30-59a", "60a>", "30-59a...
#> $ edad_asis_muna   <chr> "60a>", "60a>", "0-9a", "20-59a", "60a>", "20-59a"...
#> $ edad_estandar    <fct> 70-74, 80+, 0-4, 45-49, 60-64, 55-59, 40-44, 80+, ...